SharingService/pyproject.toml
[project]
name = "sharingservice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.121.1",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
